changed img directory for media/photos in order to render.dashobard could save images to disc and do not delete while new git push

+++ main.py
@@ -53,7 +53,10 @@ app = FastAPI()
 
 app.include_router(auth_router)
 
+RENDER_DICS = "media/photos"
+os.makedirs(RENDER_DICS,exist_ok=True)
 
+app.mount("/media",StaticFiles(directory="media"), name="media")
 
 app.mount("/static",StaticFiles(directory="static"), name="static")
 
@@ -107,16 +110,21 @@ async def add_project(request : Request,title:str = Form(...),description:str = 
     
     if empty_data == True and files == None:
         return JSONResponse({"redirect": ADMIN_PAGE_URL})
+    
     if files == None:
+        fallback_img = "media/photos/no-img.png"
         
-        target_dir = "static/img/"
-        os.makedirs(target_dir, exist_ok=True) 
-        source = "static/my-img/no-img.png"
-        destination = os.path.join(target_dir, "no-img.png")
-        
-        shutil.copy2(source, destination)
-            
-        img_path.append(destination)
+        if os.path.exists(fallback_img):
+            img_path.append(fallback_img)
+           
+        else:
+
+            target_dir = "media/photos/"
+            os.makedirs(target_dir, exist_ok=True) 
+            source = "static/my-img/no-img.png"
+            destination = os.path.join(target_dir, "no-img.png")
+            shutil.copy2(source, destination)
+            img_path.append(destination)
         
             
     else:
@@ -157,7 +165,7 @@ def delete_project(request:Request,id : int, model_type = str,admin = Depends(ge
         choosen_model = session.get(model,id)
         if choosen_model:
             for img in choosen_model.image_url:
-                if img == "static/my-img/no-img.png":
+                if img == "media/photos/no-img.png":
                     continue
                 else:
                     try:
@@ -197,7 +205,7 @@ def get_edit_content(request:Request,id:int,model_type = str,admin = Depends(get
 @app.post(ADMIN_EDIT_PROJECT,response_class = HTMLResponse)
 async def post_edit_content(request:Request,id:int,title: str = Form(...),description:str = Form(...),preview:Optional[str] = Form(None),files: List[UploadFile] = File(None),model_type = str,admin = Depends(get_current_user_from_cookies)):
     config = MODEL.get(model_type)
-    fallback_img = "static/my-img/no-img.png"
+    fallback_img = "media/photos/no-img.png"
     model = config["model"]
     markdown_text = description
     with Session(engine) as session:
@@ -240,7 +248,7 @@ async def upload_img(files,choosen_model,img_path):
     
     for file in files:
        
-        DIR = "static/img/"
+        DIR = "media/photos/"
         os.makedirs(DIR,exist_ok=True)
         location_file = os.path.join(DIR,file.filename)
